Refactor unit tests to appropriate locations
