--- docs/development/core/public/kibana-plugin-public.httpsetup.removeallinterceptors.md
@@ -1,17 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [kibana-plugin-public](./kibana-plugin-public.md) &gt; [HttpSetup](./kibana-plugin-public.httpsetup.md) &gt; [removeAllInterceptors](./kibana-plugin-public.httpsetup.removeallinterceptors.md)
-
-## HttpSetup.removeAllInterceptors() method
-
-Removes all configured interceptors.
-
-<b>Signature:</b>
-
-```typescript
-removeAllInterceptors(): void;
-```
-<b>Returns:</b>
-
-`void`
-
